Extract duplicated line function to shared helpers
