+++ packages/victory-core/src/index.js
@@ -45,3 +45,4 @@ export * as Axis from "./victory-util/axis";
 export { default as TimerContext } from "./victory-util/timer-context";
 export { default as PortalContext } from "./victory-portal/portal-context";
 export * from "./victory-util/hooks";
+export * as LineHelpers from "./victory-util/line-helpers";
